pyproject.toml
[tool.poetry]
name = "HanziLevelUp"
version = "0.6.2"
description = """
A Hanzi learning suite, with levels based on Hanzi Level Project, aka. \
nother attempt to clone WaniKani.com for Chinese.\
"""
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/patarapolw/HanziLevelUp"
homepage = "https://github.com/patarapolw/HanziLevelUp"
keywords = ['wanikani', 'chinese', 'mandarin', 'hanzi']

[tool.poetry.dependencies]
python = ">=3"
flask = "^1.0"
flask-sqlalchemy = "^2.3"
flask-migrate = "^2.2"
flask-whooshalchemy = "^0.56.0"
wordfreq = "^2.1"
requests = "^2.19"
bs4 = "^0.0.1"
jieba = "^0.39.0"
regex = "=2018.02.21"
namedlist = "^1.7"
python-dateutil = "^2.7"
pyexcel = "^0.5.8"
#pyexcel-xlsxwx = "^0.1.7"
pyexcel-xlsxwx = {path = "../pyexcel-xlsxwx"}

[tool.poetry.dev-dependencies]
#pyopenssl = "^18.0"
sqlite-web = {git = "https://github.com/patarapolw/sqlite-web.git", branch = "master"}
jupyter = "^1.0"
notebook = "^5.6"
simplecel = {path = "../simplecel"}
pyexcel-xls = "^0.5.7"
